+++ src/main/java/com/sporthustle/hustle/src/user/model/LoginRes.java
@@ -1,6 +1,6 @@
 package com.sporthustle.hustle.src.user.model;
 
-import com.sporthustle.hustle.common.security.model.TokenInfo;
+import com.sporthustle.hustle.common.jwt.model.TokenInfo;
 import lombok.Builder;
 import lombok.Getter;
